+++ src/versioning.rs
@@ -1,9 +1,9 @@
-use std::collections::HashMap;
+use std::{collections::HashMap, env};
 
 use base64::prelude::*;
 use chrono::Utc;
 use node_semver::Version;
-use serde::{Deserialize, Serialize};
+use serde::{Deserialize, Deserializer, Serialize, Serializer, ser::SerializeSeq};
 use tracing::{error, warn};
 
 #[derive(Debug, Default, Serialize, Deserialize)]
@@ -12,10 +12,35 @@ pub struct Versioning {
     build_time: String,
     built_with: BuiltWith,
     repository: Repository,
-    sources: Vec<Source>,
+    #[serde(
+        serialize_with = "sources_serialize",
+        deserialize_with = "sources_deserialize"
+    )]
+    sources: HashMap<String, Source>,
 }
 
-#[derive(Debug, Default, Serialize, Deserialize)]
+fn sources_serialize<S>(sources: &HashMap<String, Source>, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    let mut seq = serializer.serialize_seq(Some(sources.len()))?;
+    for source in sources.values() {
+        seq.serialize_element(&source)?;
+    }
+    seq.end()
+}
+
+fn sources_deserialize<'de, D>(deserializer: D) -> Result<HashMap<String, Source>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    Ok(Vec::<Source>::deserialize(deserializer)?
+        .into_iter()
+        .map(|v| (v.id.clone(), v))
+        .collect())
+}
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct BuiltWith {
     toolchain: String,
@@ -67,7 +92,32 @@ pub struct Developers {
     github: Option<String>,
 }
 
-type UpdatedExtensions = Vec<(String, HashMap<String, String>)>;
+#[derive(Debug, Default, Serialize, Deserialize)]
+pub struct Metadata {
+    #[serde(flatten)]
+    repositories: HashMap<String, MetadataRepository>,
+}
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
+pub struct MetadataRepository {
+    #[serde(flatten)]
+    extensions: HashMap<String, MetadataExtension>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct MetadataExtension {
+    build_time: String,
+    built_with: BuiltWith,
+}
+
+pub type UpdatedExtensions = Vec<(String, UpdateTypes, HashMap<String, String>)>;
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum UpdateTypes {
+    Addition,
+    Update,
+    Deletion,
+}
 
 impl Versioning {
     pub fn new(response_content: &str) -> Result<Versioning, ()> {
@@ -107,10 +157,9 @@ impl Versioning {
 
     pub fn update(
         &mut self,
-        repository_versioning: &Versioning,
-    ) -> Result<UpdatedExtensions, bool> {
-        let mut updated_extensions = vec![];
-
+        metadata: &mut Metadata,
+        repository_versioning: Versioning,
+    ) -> Result<UpdatedExtensions, ()> {
         if self
             .built_with
             .types
@@ -120,49 +169,124 @@ impl Versioning {
                 .built_with
                 .types
                 .parse::<Version>()
-                .unwrap_or_else(|_| Version::parse("0.0.0").unwrap())
+                .unwrap()
         {
             error!(
-                "The repository was build with an invalid @paperback/types version {}, expected version {} or higher",
+                "The repository was build with a @paperback/types version {} which was too low, expected version {} or higher",
                 repository_versioning.built_with.types, self.built_with.types
             );
-            return Err(true);
+            return Err(());
         }
 
-        for repository_extension in repository_versioning.sources.iter() {
-            let mut new = true;
+        let mut updated_extensions = vec![];
 
-            for (index, registry_extension) in self.sources.iter().enumerate() {
-                if repository_extension.id != registry_extension.id {
-                    continue;
+        let mut shared_extensions = vec![];
+
+        let mut repository_extensions = repository_versioning
+            .sources
+            .keys()
+            .cloned()
+            .collect::<Vec<String>>();
+
+        let mut registry_extensions = metadata
+            .repositories
+            .get(&env::var("REPOSITORY").unwrap())
+            .cloned()
+            .unwrap_or_default()
+            .extensions
+            .keys()
+            .cloned()
+            .collect::<Vec<String>>();
+
+        for repository_index in (0..repository_extensions.len()).rev() {
+            for registry_index in (0..registry_extensions.len()).rev() {
+                if repository_extensions[repository_index] == registry_extensions[registry_index] {
+                    shared_extensions.push(repository_extensions[repository_index].clone());
+                    repository_extensions.remove(repository_index);
+                    registry_extensions.remove(registry_index);
+
+                    break;
                 }
+            }
+        }
 
-                new = false;
+        metadata
+            .repositories
+            .entry(env::var("REPOSITORY").unwrap())
+            .or_insert_with(|| MetadataRepository {
+                extensions: HashMap::new(),
+            });
 
-                if repository_extension
+        for extension in repository_extensions.iter() {
+            self.sources.insert(
+                extension.to_string(),
+                repository_versioning
+                    .sources
+                    .get(extension)
+                    .unwrap()
+                    .clone(),
+            );
+
+            metadata
+                .repositories
+                .get_mut(&env::var("REPOSITORY").unwrap())
+                .unwrap()
+                .extensions
+                .insert(
+                    extension.to_string(),
+                    MetadataExtension {
+                        build_time: repository_versioning.build_time.clone(),
+                        built_with: repository_versioning.built_with.clone(),
+                    },
+                );
+
+            updated_extensions.push((extension.clone(), UpdateTypes::Addition, HashMap::new()));
+        }
+
+        for extension in shared_extensions.iter() {
+            if repository_versioning
+                .sources
+                .get(extension)
+                .unwrap()
+                .version
+                .parse::<Version>()
+                .unwrap_or_else(|_| Version::parse("0.0.0").unwrap())
+                > self
+                    .sources
+                    .get(extension)
+                    .unwrap()
                     .version
                     .parse::<Version>()
-                    .unwrap_or_else(|_| Version::parse("0.0.0").unwrap())
-                    > registry_extension.version.parse::<Version>().unwrap()
-                {
-                    self.sources[index] = repository_extension.clone();
+                    .unwrap()
+            {
+                self.sources.insert(
+                    extension.to_string(),
+                    repository_versioning
+                        .sources
+                        .get(extension)
+                        .unwrap()
+                        .clone(),
+                );
 
-                    updated_extensions.push((repository_extension.id.clone(), HashMap::new()));
-                }
-
-                break;
+                updated_extensions.push((extension.clone(), UpdateTypes::Update, HashMap::new()));
             }
+        }
 
-            if new {
-                self.sources.push(repository_extension.clone());
+        for extension in registry_extensions.iter() {
+            self.sources.remove(extension);
 
-                updated_extensions.push((repository_extension.id.clone(), HashMap::new()));
-            }
+            metadata
+                .repositories
+                .get_mut(&env::var("REPOSITORY").unwrap())
+                .unwrap()
+                .extensions
+                .remove(extension);
+
+            updated_extensions.push((extension.clone(), UpdateTypes::Deletion, HashMap::new()));
         }
 
         if updated_extensions.is_empty() {
             warn!("There are no extensions to update");
-            return Err(false);
         }
 
         self.build_time = Utc::now().format("%Y-%m-%dT%H:%M:%S%.3fZ").to_string();
@@ -181,5 +305,42 @@ impl Versioning {
         }
 
         Ok(updated_extensions)
+    }
+}
+
+impl Metadata {
+    pub fn new(response_content: &str) -> Result<Metadata, ()> {
+        match BASE64_STANDARD.decode(response_content.replace("\n", "")) {
+            Ok(bytes) => match serde_json::from_slice(bytes.as_slice()) {
+                Ok(metadata) => Ok(metadata),
+                Err(err) => {
+                    error!(
+                        "An error occurred while deserializing the response content to JSON: {}",
+                        &err
+                    );
+                    Err(())
+                }
+            },
+            Err(err) => {
+                error!(
+                    "An error occurred while base64 decoding the response content: {}",
+                    &err
+                );
+                Err(())
+            }
+        }
+    }
+
+    pub fn to_base64(&self) -> Result<String, ()> {
+        match serde_json::to_string(&self) {
+            Ok(metadata_string) => Ok(BASE64_STANDARD.encode(&metadata_string)),
+            Err(err) => {
+                error!(
+                    "An error occurred while encoding the metadata file into base64: {}",
+                    &err
+                );
+                Err(())
+            }
+        }
     }
 }