+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 - Ensure a clean exit (exit code 0) when no extensions need updating, display a warning instead (@Celarye)
+- Added extension deletion support by rewriting the versioning update logic (@Celarye)
 
 ### Bug fixes
 - Extension duplication (@Celarye)